requirements.txt
# python_version: '3.8.5'

discord.py==1.6.0
python-dotenv==0.15.0
